styling poems, placing content into centered div
This commit inserts some inline css (meaning css within the html file)
to stye the content we’ve added in the first few commits.  You’ll
notice that inline css is included in the <head> tag using a <style>
tag.  Eventually, we will break the css out into its own file, but for
now, we’ll mess around with it inside the html file.

Some things to take note of are:

+ The CSS applied to “#content” is being applied to whichever element
has the id of “content.”  In this case, we’ve styled the #content
wrapper to have a width of 45% of the div it is contained within, in
this case the BODY tag, and centered it.

+ The CSS applied to “.pome” is being applied to every element with the
class “pome.”  This CSS is telling the browser to respect all white
space so that the spacing and line breaks of the text within those divs
appear correctly on screen.

+ The CSS applied to “audio, img” set the width of all audio and img
elements to be 100% of the width of their containing div.

+ We also linked to a Google Font called Arvo and applied them to the
titles on the page, just for style.