includes google map for geo-based navigation
This commit adds a map at the top of the site, which we will style and
change to use as a geography-based way to navigate through the poems
and other media in the site to help more visually represent the trips
enumerated on the navigation bar we just added.

Notice that in order to add this map, we added a link to google maps in
the <head> tag, added a div with the id “map-canvas,” added some CSS
styling to set the size of the map, and most importantly added our
first bits of Javascript in an analogous way to how we added inline CSS
but using a <script> tag rather than a <style> tag.

You can read more about using the google maps api with javascript here:
 https://developers.google.com/maps/documentation/javascript/

Note:  If you want to use Google Maps in your own project, you will
need to sign up as a developer with google to get your own API key.
You can read more about it at the above link.