creates sticky navbar
This function, run whenever the mouse is scrolled, makes the navigation
bar “stick” to the top of the screen when the site is scrolled down.
It does this by checking the position of the #map-canvas div.  When it
is has been scrolled off the screen, the CSS position of the navbar is
reset to be “fixed” at “top:0” (e.g. the top of the screen).