creating unique links to each poem
When we first added the poem content, each .pome-block had a
strange-looking ID attached to it.  These strings were created as a
shorthand way of encoding information about when the poems were written
and by whom based on the metadata from the emails they were originally
sent in.  Because they include a timestamp, they will be unique IDs, so
this function, which waits to run until the page has loaded, sets the
currently empty href in the <a href=“#”> tags of each poem to be <a
href=“#uniqueID”>

Because this string will be used again (in the next commit!) to
generate by lines (e.g. this poem was written at THIS TIME by THIS
PERSON), we are also going to store the individual elements of each
string in a legible way as “data-attributes” of each .pome-block to
easily access later.

Each poem’s unique string is formatted in the style
YEAR-MONTH-DAY-TIME-TRIP-AUTHOR, so for each pome-block we create
data-attributes to store the related values.  So, the first poem’s ID
“2014-07-18-0919-DETROIT-KP” will be stored within the .pome-class div
as <div class=“pome-block” data-year=“2014” data-month=“07”
data-day=“18” data-time=“0919” data-trip=“DETROIT” data-author=“KP”>