adding bylines to each poem
This commit adds two functions to the javascript file which run when
the page is loaded and insert bylines after each poem title.

The first accesses the data-attributes we created for each poem in the
last commit and concatenates them into a human-readable phrase telling
readers when and by whom each poem was written.  In order to turn the
“data-time” attribute into a clock-like style, we used regular
expressions, which you can read more about here:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Ex
pressions

The second iterates through the page and inserts these bylines after
every “h2” element, which we have organized to be poem titles.