changes background images on scroll
Thus far, the background of the site has been static.  In this commit,
we’ve hardcoded links to a collection of images associated which each
trip which we will rotate through as the user scrolls.  When the
content from one trip ends and a new trip’s content begins, the
background image will change to a picture associated with the current
trip.

In the function change_bgimgs(), which runs whenever the mouse is
scrolled, we iterate through each trip to check if the top of its
content is in the top quarter of the screen or off the top of the
screen AND that the bottom of its content is in the bottom quarter of
the screen or off the bottom of the screen.  If both these things are
true, then we know the current trip is the one being displayed on the
screen.  We then check to see if the current background image is
associated with the current trip.  If yes, we leave the image as is.
If no, we change the background image to a random image from the
collection associated with the current trip.

We also added some CSS to the make the transition between background
images smoother.