changing background image via navbar
At this point, the site is mostly done, but one frustrating behavior is
that the background image only changes when the user is scrolling.  So,
if someone were to notice the changing background and want to just see
the variety of pictures, they’d have to scroll back and forth at a trip
transition point over and over.  That’s frustrating, so this last bit
of code adds an event listener to the navbar buttons so that when the
same button is clicked repeatedly, a random picture associated with
that trip will be chosen and loaded as the background-image.  This
mostly works, though it does mean that sometimes the random picture
chosen will be the same as the one currently on screen, so there will
appear to be no response.