hide current splash and create alt-splash div
Next, we are going to add a second way of navigating the site.  Rather
than just moving through the content by location—through the map and
navbar—we will build a “physical interface” using the computer’s
built-in camera to recognize and read cards or tokens we make which
index into the content of the site in a less direct, more whimsical way.

To start, we will hide the #splash div and create an #alt-splash div
where we will mess around with building this new interface.