taking still photos from built-in video feed
First, we are going to hook up the camera and get it working.  I used
code shared in this tutorial on how to access the computer’s built-in
camera using the browser and made changes to get it working in this
context.

https://developer.mozilla.org/en-US/docs/Web/API/WebRTC_API/Taking_still
_photos

In order to get this working, I needed to move the link to the external
js file (pomes.js) to the bottom of the html page because the new code
accessing the camera was being run before some of the html elements
were loaded.

With the link moved, the window.onload function in the js file is no
longer necessary, so I also deleted that.

You an read a little more about this here:
http://stackoverflow.com/questions/15751362/window-onload-in-external-sc
ript-gets-ignored-in-javascript

If you look at this commit in the browser, you should first get an
alert asking you for permission to let the browser access your
computer’s camera (click “Allow”).  Give the page a couple seconds to
load and you should see an active video feed in the upper left hand
corner of the page.  When you click the “Take Photo” button, you should
see a still photo displayed on the right side of the screen.  Check it
out!  Take some pictures!