styling the alt-splash div
Just so things look nice while we’re messing around, we’ve added some
styling to the new #alt-splash.  In particular, you’ll notice:

+ We set a default background to the #video and #canvas elements.  This
is so viewers know they should expect content to be there, in case the
video feed or still photo displays aren’t working.
+ For similar reasons, we included text over the video and still
pictures explaining what content should be displayed in the blank div.
The CSS needed to place text over the video and canvas elements are
called out with comments in the CSS, just in case you want to do this
in a future project.
+ Finally, you’ll also notice we switched the id the sticky navbar is
looking at to decide whether to fix itself to the top of the screen or
not.  Since we are just messing with #alt-splash right now, this will
work, but we’ll have to rewrite the sticky nav code to be responsive
when we want users to be able to choose which interface they are
viewing.