calculating the average color of an image
Now that we can take a picture, we need to decide how we want to use
those pictures to create an image-based interface to the site’s poems
and other media.

We are going to write code to recognize tokens based on color.  This
will allow us the flexibility to create ad hoc, DIY tokens made of any
material, displaying any shape, as long as they can be photographed and
produce an image which is mostly one color.  On the other hand, it
means we will be relying on color recognition and therefore that the
shadow and lighting of our images will impact our ability to recognize
the tokens we photograph.

As a first step, we wrote a function that takes an image and calculates
the average color of that image by creating an array of the image’s
pixels, iterating over that array, averaging the rgb values of every
5th pixel (to save time), and returning that average rgb.

We based our code on this average color function:
https://gist.github.com/olvado/1048628

Because we are displaying the images we grab from our video feed in a
canvas object, we rewrote the example code to take a canvas rather than
an image object.  In addition, we’ve added an event listener to the
“Take Photo” button so that the average color of the photo taken will
be logged to the console each time you click the button.  Take a look
and mess around to see if it behaves how you’d expect.

In case you are interested in other types of image analysis and
recognition, you might also look into:

+ OCR (or text recognition: http://kdzwinel.github.io/JS-OCR-demo/)
+ shape recognition
(http://kevinmehall.net/2009/line-and-shape-recognition,
http://katspaugh.github.io/whiteboard/examples/brain/#triangle, or
http://techslides.com/object-detection-with-html5-getusermedia are all
interesting resources)
+ face recognition
(http://stackoverflow.com/questions/7291065/any-library-for-face-recogni
tion-in-javascript)