assigning color name to avg RGB value from pics
Now, using the avgColor and distanceBetween functions from the last two
commits, we are going to write a function that can identify the color
name that best describes the average RGB value calculated from an image.

First, we create a dictionary mapping all of the “pure” color names to
their RGB values.  These are the colors against which we are going to
test our RGB values (e.g. the “corner” colors, per the RGB cube model
described in the last commit).

Next, we iterate through these test colors, creating a second
dictionary mapping the distance between the average color of a given
image and a test color to the name of the test color.

We then find the minimum distance calculated and return the color name
associated with that distance.

We also changed the eventListener on the “Take Photo” so that it now
log the calculated color name to the console.

Note:  If you are getting “black” for a lot of your images, make sure
your space is well lit enough for the camera to capture color
information.