navigating through the page using pictures
Now that we have access to the color name of each picture as we take
them, we have to somehow index those names into the contents of the
site.  In this commit, we’ve placed color name IDs, matching the “pure”
or “corner” colors we’ve written our RGB code around, as anchor links
throughout the site and written a few lines of code, at the bottom of
pomes.js, to scroll the page to the location of the anchor link
matching the color calculated and identified from a given picture.

So, when you click the “Take Photo” button in this commit, it should
automatically send you to the place in the site with the anchor link
that matches the color of the token you photographed.