adding button to toggle btw #splash & #alt-splash
This commit adds a button to the top of the site which allows the
viewer to toggle between the original, map splash area (which is the
default to display) and the alternative, camera-navigated splash area.

This is done by creating a button which displays atop either splash
area and adding an eventListener to this button which checks the
“display” property of each splash area and, depending on which is
currently being displayed, toggles to the other splash area when the
button is clicked.