fixing sticky nav bar
If you played around with the last commit, you probably noticed that
the sticky navbar is broken whenever the map splash area is displayed.
In this commit, we fixed this by updating the make_navbar_sticky
function check the position of BOTH splash areas so that, if one of
them is both being displayed and has been scrolled off the visible
screen, the navbar will affix itself to the top of the page.