fixing the "To Map" button
Another piece that broke when we added the toggle button was the “Top
Map” button.  For one, it no longer makes sense to have a “To Map”
button when the camera splash is being displayed.  But we also need to
change the anchor link to which that navigation link directs us based
on which splash screen is being displayed.

Because we already wrote out the logic to identify which splash is
currently being displayed in the function toggleSplash, we can just
redefine the a.href of that navigation link within that function.